refactor(4a): move snapshot I/O helpers to daemon/snapshot-gather.ts
